app/requirements.txt
Flask==2.3.0
Werkzeug==2.3.0
numpy>=1.24.0
Pillow>=9.5.0
opencv-contrib-python>=4.8.0
scikit-image>=0.20.0
scipy>=1.11.0
tensorflow>=2.13.0
keras>=2.13.0
tensorflow>=2.10.0
keras>=3.0.0
opencv-python>=4.5.0
pillow>=9.0.0
numpy>=1.21.0
flask>=2.0.0
